# api_spec.yaml
openapi: 3.0.0
info:
  title: Sample Inventory API
  version: 1.0.0
  description: |
    A simple service for managing inventory items. The API supports creating,
    listing, updating and deleting items and is useful for testing typical
    business workflows.
paths:
  /ping:
    get:
      summary: Health check
      operationId: controllers.ping
      responses:
        '200':
          description: Returns pong message
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: pong
  /items:
    get:
      summary: List all items
      operationId: controllers.list_items
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Filter results by case-insensitive substring match on the item name. Returns all items when omitted.
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedItems'
    post:
      summary: Create a new item
      operationId: controllers.create_item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '201':
          description: Created item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
  /items/low-stock:
    get:
      summary: List items below a quantity threshold
      operationId: controllers.low_stock
      parameters:
        - in: query
          name: threshold
          required: true
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - in: query
          name: per_page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
      responses:
        '200':
          description: Items with quantity below threshold
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedItems'
  /items/{id}:
    get:
      summary: Get item by ID
      operationId: controllers.get_item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Not found
    put:
      summary: Update item
      operationId: controllers.update_item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemInput'
      responses:
        '200':
          description: Updated item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Not found
    delete:
      summary: Delete item
      operationId: controllers.delete_item
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Not found
components:
  schemas:
    PaginatedItems:
      type: object
      description: Paginated response for item listings
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
        page:
          type: integer
          description: Current page number
          example: 1
        per_page:
          type: integer
          description: Number of items per page
          example: 10
        total:
          type: integer
          description: Total number of items
          example: 100
        pages:
          type: integer
          description: Total number of pages
          example: 10
        has_next:
          type: boolean
          description: Whether there is a next page
          example: true
        has_prev:
          type: boolean
          description: Whether there is a previous page
          example: false
    Item:
      type: object
      description: An inventory record identified by an integer ID, a name, a quantity and a price.
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: example item
        quantity:
          type: integer
          example: 10
        price:
          type: number
          example: 9.99
    ItemInput:
      type: object
      description: Payload used when creating or updating an inventory item.
      required:
        - name
        - quantity
        - price
      properties:
        name:
          type: string
          example: example item
        quantity:
          type: integer
          example: 10
        price:
          type: number
          example: 9.99